macOS Phase 0: cross-compile baseline + CI matrix
- libc-shim: kqueue-backed epoll (epoll_kqueue.rs), host errno abstraction,
  mac variants for stat/pthread/sendfile/getrandom/sem_timedwait/mallinfo
- linker: self-defined bionic-layout dl_phdr_info (no host ELF types)
- client: target-gated build.rs dylibs + x11/cpal deps; macOS stub backends
  under platform/macos/{eglut,gamepad,game_window,audio} keep darwin compiling
- CI: .github/workflows/macos.yml (macos-15 arm64 + macos-13 Intel) with
  Linux regression job
- docs/PORT_MACOS.md: full port plan; AGENTS.md pointer added

cargo check green on aarch64/x86_64-apple-darwin and Linux

+++ crates/client/src/jnivm_class_wrappers.rs
@@ -99,17 +99,17 @@ fn statvfs_space(path: *const i8, field_fn: fn(&libc::statvfs) -> u64) -> jlong 
 
 unsafe extern "C" fn File_getTotalSpace(_env: *mut JNIEnv, self_: jobject) -> jlong {
     let f = self_ as *const FileObject;
-    statvfs_space((*f).path.as_ptr(), |s| s.f_blocks)
+    statvfs_space((*f).path.as_ptr(), |s| s.f_blocks as u64)
 }
 
 unsafe extern "C" fn File_getUsableSpace(_env: *mut JNIEnv, self_: jobject) -> jlong {
     let f = self_ as *const FileObject;
-    statvfs_space((*f).path.as_ptr(), |s| s.f_bavail)
+    statvfs_space((*f).path.as_ptr(), |s| s.f_bavail as u64)
 }
 
 unsafe extern "C" fn File_getFreeSpace(_env: *mut JNIEnv, self_: jobject) -> jlong {
     let f = self_ as *const FileObject;
-    statvfs_space((*f).path.as_ptr(), |s| s.f_bfree)
+    statvfs_space((*f).path.as_ptr(), |s| s.f_bfree as u64)
 }
 
 fn register_file_class(env: *mut JNIEnv) {
